# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Robotic Workbench",
    "version": "0.1.0"
  },
  "paths": {
...
    "/vision/observations": {
      "get": {
        "summary": "Observations",
        "operationId": "vision_observations_observations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Observations"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/vision/preview": {
      "get": {
        "summary": "Preview",
        "operationId": "vision_preview_preview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/vision/mock/scene": {
      "put": {
        "summary": "Scene",
        "operationId": "vision_scene_mock_scene_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MockScene"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MockScene"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/vision/mock/reset": {
      "post": {
        "summary": "Reset",
        "operationId": "vision_reset_mock_reset_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MockScene"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Error"
      },
...
      "MockScene": {
        "properties": {
          "tools": {
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "type": "array",
            "title": "Tools"
          },
          "tray_clear": {
            "type": "boolean",
            "title": "Tray Clear",
            "default": true
          },
          "camera_ok": {
            "type": "boolean",
            "title": "Camera Ok",
            "default": true
          },
          "age_seconds": {
            "type": "number",
            "maximum": 3600.0,
            "minimum": 0.0,
            "title": "Age Seconds",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tools"
        ],
        "title": "MockScene"
      },
      "Observation": {
        "properties": {
          "tool_id": {
            "type": "string",
            "title": "Tool Id"
          },
          "location": {
            "type": "string",
            "enum": [
              "source",
              "tray",
              "missing",
              "uncertain"
            ],
            "title": "Location"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tool_id",
          "location",
          "confidence"
        ],
        "title": "Observation"
      },
      "Observations": {
        "properties": {
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "type": "array",
            "title": "Tools"
          },
          "tray_clear": {
            "type": "boolean",
            "title": "Tray Clear"
          },
          "camera_ok": {
            "type": "boolean",
            "title": "Camera Ok"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "captured_at",
          "tools",
          "tray_clear",
          "camera_ok"
        ],
        "title": "Observations"
      },
...
    }
  }
}